% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdarr_standard.R
\name{sdar}
\alias{sdar}
\title{SDAR-algorithm}
\usage{
sdar(data, x, y, verbose = TRUE, plot = TRUE, plotFun = FALSE, ...)
}
\arguments{
\item{data}{Data record to analyze. Labels of the data columns will be used
as units.}

\item{x, y}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns with x and y
within data.}

\item{verbose, plot}{Give a summarizing report / show a plot of the final fit.}

\item{plotFun}{Set to \code{TRUE} to get a plot-function for the final fit with
the results for later use.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Pass parameters to downstream
functions: set \code{verbose.all}, \code{plot.all} and \code{plotFun.all} to \code{TRUE} to get
additional diagnostic information during processing data.}
}
\value{
A list containing a data.frame with the results of the final fit,
lists with the quality- and fit-metrics, and a list containing the crated
plot-function(s) (if \code{plotFun = TRUE}).
}
\description{
Run the SDAR algorithm as standardized in ASTM E3076-18. Will
use numerous linear regressions (\code{.lm.fit()} from the stats-package) and
can be painfully slow for test data with high resolution. See the article
\href{https://soylentorange.github.io/sdarr/articles/speed_improvment.html}{Speed Benchmarking the SDAR-algorithm}
for further information.
}
\note{
The function can use parallel processing via the
\href{https://furrr.futureverse.org/}{furrr-package}. To use this feature, set
up a plan other than the default sequential strategy beforehand.
}
\examples{
# Synthesize a test record resembling EN AW-6060-T66
# Explicitly set names to "strain" and "stress",
# set effective number of bits in the x-data to 12
# to limit the number of data points.

Al_6060_T66 <- synthesize_test_data(
  slope = 69000,
  yield.y = 160,
  ultimate.y = 215,
  ultimate.x = 0.08,
  x.name = "strain",
  y.name = "stress",
  toe.start.y = 3, toe.end.y = 10,
  toe.start.slope = 13600,
  enob.x = 12
)


# use sdar() to analyze the synthetic test record
# will print a report and give a plot of the final fit
\donttest{
result <- sdar(Al_6060_T66, strain, stress)
}

}
\references{
Lucon, E. (2019). Use and validation of the slope determination
by the analysis of residuals (SDAR) algorithm (NIST TN 2050; p. NIST TN
2050). National Institute of Standards and Technology.
https://doi.org/10.6028/NIST.TN.2050

Standard Practice for Determination of the Slope in the Linear
Region of a Test Record (ASTM E3076-18). (2018).
https://doi.org/10.1520/E3076-18

Graham, S., & Adler, M. (2011). Determining the Slope and Quality
of Fit for the Linear Part of a Test Record. Journal of Testing and
Evaluation - J TEST EVAL, 39. https://doi.org/10.1520/JTE103038
}
\seealso{
\code{\link[=sdar_lazy]{sdar_lazy()}} for the random sub-sampling modification of the
SDAR-algorithm.
}
