% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdarr-package.R
\docType{package}
\name{sdarr-package}
\alias{sdarr}
\alias{sdarr-package}
\title{sdarr: Slope Determination by Analysis of Residuals}
\description{
Slope Determination by Analysis of Residuals (SDAR) for automated and objective linear-fitting of mechanical test-data (see: standard ASTM E3076-18 \doi{10.1520/E3076-18} or a detailed description in E. Lucon (2019) \doi{10.6028/NIST.TN.2050}). As the SDAR-algorithm heavily uses linear regressions, a faster version is additionally available to find the optimum region for the final linear regression by random sub-sampling within the normalized range of the test-data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://soylentorange.github.io/sdarr/}
  \item Report bugs at \url{https://github.com/soylentOrange/sdarr/issues}
}

}
\author{
\strong{Maintainer}: Robert Wendlandt \email{robert.wendlandt@uksh.de} (\href{https://orcid.org/0000-0001-6368-1622}{ORCID})

Other contributors:
\itemize{
  \item UKSH [copyright holder, funder]
}

}
\keyword{internal}
