% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdarr.R
\name{sdarr}
\alias{sdarr}
\title{sdarr: SDAR-algorithm}
\usage{
sdarr(data, x, y, verbose = "report", showPlots = "report", savePlots = F)
}
\arguments{
\item{data}{Data record to analyze. Labels of the data columns will be used
as units.}

\item{x}{Use tidy selection to specify x within the data.}

\item{y}{Use tidy selection to specify y within the data.}

\item{verbose}{Give informational messages during computation defaults to
"report" to only show a summarizing information set to "all" to also give
messages from the individual steps set to "none" to be quiet. Can be
abbreviated.}

\item{showPlots}{Show plots during computation defaults to "report" to only
show the plot of the final fit set to "all" to also show plots from the
individual steps set to "none" to be quiet. Can be abbreviated.}

\item{savePlots}{Give plot functions with the result.}
}
\value{
A list containing a data-frame with the results of the final fit, a
  list with the quality- and fit-metrics, and a list containing the crated
  plot-functions (if savePlots was set to TRUE).
}
\description{
Run the SDAR algorithm as standardized in ASTM E3076-18. Will
  use numerous linear regressions (.lm.fit form the stats-package) and can be
  painfully slow for test data with high resolution.
}
\note{
The function will use parallel processing via the furrr-package. To use
  this feature, set up a plan other than the default sequential strategy
  beforehand.
}
\references{
Lucon, E. (2019). Use and validation of the slope determination
  by the analysis of residuals (SDAR) algorithm (NIST TN 2050; p. NIST TN
  2050). National Institute of Standards and Technology.
  https://doi.org/10.6028/NIST.TN.2050

Standard Practice for Determination of the Slope in the Linear
  Region of a Test Record (ASTM E3076-18). (2018).
  https://doi.org/10.1520/E3076-18

Graham, S., & Adler, M. (2011). Determining the Slope and Quality
  of Fit for the Linear Part of a Test Record. Journal of Testing and
  Evaluation - J TEST EVAL, 39. https://doi.org/10.1520/JTE103038
}
