% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesizeTestData.R
\name{synthesize_test_data}
\alias{synthesize_test_data}
\title{Synthesize Test Data}
\usage{
synthesize_test_data(
  slope,
  yield.y,
  yield.xp = 0.002,
  ultimate.y,
  ultimate.x,
  offset = 0,
  toe.initial.y = 0,
  toe.initial.slope = slope,
  toe.max.y = 0,
  enob.x = 14,
  enob.y = 14,
  enob.x_FS = 16,
  enob.y_FS = 16,
  enob.x_noise = 0,
  enob.y_noise = 0,
  x.name = "strain",
  y.name = "stress",
  x.unit = NULL,
  y.unit = "MPa"
)
}
\arguments{
\item{slope}{Slope in the linear region.}

\item{yield.y}{y-value at yield point.}

\item{yield.xp}{Plastic deformation at yield point. Defaults to 0.002.}

\item{ultimate.y}{Maximum y-value in the post-linear region.}

\item{ultimate.x}{Maximum x-value in the post-linear region.}

\item{offset}{Value of y-offset.}

\item{toe.initial.y}{Intersection of toe-region with y-axis (before adding an
offset).}

\item{toe.initial.slope}{Initial slope of toe region.}

\item{toe.max.y}{End of toe region.}

\item{enob.x}{Effective number of bits for the synthetic data x-range. Will
determine the number of points in the returned data (i.e. 2^enob.x). Also
used for adding quantization noise.}

\item{enob.y}{Effective number of bits for the synthetic data y-range. Used
for adding quantization noise.}

\item{enob.x_FS}{Effective number of bits for the full-scale x-range. Using
to determine level of quantization/data-noise in x-values.}

\item{enob.y_FS}{Effective number of bits for the full-scale y-range. Using
to determine level of quantization/data-noise in y-values.}

\item{enob.x_noise}{Add noise to x-data. Give the effective number of bits
for the full-scale x-range.}

\item{enob.y_noise}{Add noise to y-data. Give the effective number of bits
for the full-scale y-range.}

\item{x.name}{Name for x-values. Defaults to \code{"strain"}.}

\item{y.name}{Name for y-values. Defaults to \code{"stress"}.}

\item{x.unit}{Unit for x-values. Can be \code{NULL}.}

\item{y.unit}{Unit for y-values. Can be \code{NULL}. Defaults to \code{"MPa"}.}
}
\value{
A data.frame with the synthetic data. Units (when provided) are
given as variable labels.
}
\description{
Synthesize Test Data by Ramberg-Osgood equation.
}
\details{
Use the Ramberg-Osgood equation in the formulation of Hill to create
synthetic data similar to quasi-static mechanical testing. Parameters K and
n of the Ramberg-Osgood equation are calculated from the yield-strength
(typically at 0.2\% offset) and a second point in the post-yield region
(e.g. the point of ultimate stress) as proposed by Hill.
}
\note{
As random values are drawn, set a random seed (\code{\link[=set.seed]{set.seed()}}) beforehand
to get reproducible results.
}
\examples{
# Synthesize a test record resembling Al 6060 T66
# (Values according to Metallic Material Properties Development and
# Standardization (MMPDS) Handbook)
Al_6060_T66 <- synthesize_test_data(slope = 68000,
                                    yield.y = 160,
                                    ultimate.y = 215,
                                    ultimate.x = 0.091)

plot(x = Al_6060_T66$strain, y = Al_6060_T66$stress,
     type = "l",
     xlab = "strain", ylab = "stress (in MPa)")

}
\references{
Hill, H. N. (1944). Determination of stress-strain relations
from" offset" yield strength values. Aluminum Co of America Pittsburgh Pa.

Ramberg, W., & Osgood, W. R. (1943). Description of Stress-Strain
Curves by Three Parameters; National Advisory Committee for Aeronautics
Technical Note. NACA-TN-902.

Rice, R. C., Jackson, J. L., Bakuckas, J., & Thompson, S. (2003).
Metallic Materials Properties Development and Standardization (MMPDS)
(DOT/FAA/AR-MMPDS-01).
}
