% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdarr-package.R
\docType{package}
\name{sdarr-package}
\alias{sdarr-package}
\alias{_PACKAGE}
\title{sdarr: Slope Determination by Analysis of Residuals}
\description{
Implementation of the SDAR algorithm for Slope Determination by Analysis of Residuals as standardized in ASTM E3076-18. It allows for automated and objective linear-fitting of mechanical test-data. As the algorithm heavily uses linear regressions, a faster version is also available to find the optimum region for linear fitting by random sub-sampling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://soylentorange.github.io/sdarr/}
  \item Report bugs at \url{https://github.com/soylentOrange/sdarr/issues}
}

}
\author{
\strong{Maintainer}: Robert Wendlandt \email{robert.wendlandt@uksh.de} (\href{https://orcid.org/0000-0001-6368-1622}{ORCID})

}
\keyword{internal}
