% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdarrLazy.R
\name{sdarr.lazy}
\alias{sdarr.lazy}
\title{Random sub-sampling variant of the SDAR-algorithm}
\usage{
sdarr.lazy(
  data,
  x,
  y,
  fit.rep = 5,
  fit.candidates = 20,
  cutoff_probability = 0.975,
  quality_penalty = 0.1,
  enforce_subsampling = FALSE,
  verbose = "report",
  showPlots = "report",
  savePlots = FALSE
)
}
\arguments{
\item{data}{Data record to analyze. Labels of the data columns will be used
as units.}

\item{x, y}{Use <\code{\link[dplyr:dplyr_tidy_select]{tidy-selections}}> to specify x
and y within the data record.}

\item{fit.rep}{Repetitions of random sub-sampling and fitting.}

\item{fit.candidates}{Give a number for selecting optimum fit candidates
(ordered decreasingly by normalized residuals) for each of the repetitions
of random sub-sampling.}

\item{cutoff_probability}{Cut-off probability for estimating optimum size of
sub-sampled data range via logistic regression.}

\item{quality_penalty}{Factor to down-weight fits with inferior data- and
fit-quality metrics.}

\item{enforce_subsampling}{Set to \code{TRUE}, to use sub-sampling method even
when it is computationally more expensive than the standard SDAR-algorithm.}

\item{verbose, showPlots}{Give informational messages and plots during
computation. Defaults to \code{"report"} to only show a report and a plot of the
final fit. Set to \code{"all"} to also give messages from the individual steps.
Set to \code{"none"} to be quiet. Can be abbreviated.}

\item{savePlots}{Set to \code{TRUE} to get plot-functions with the result for
later use.}
}
\value{
A list containing a data-frame with the results of the final fit, a
list with the quality- and fit-metrics, and a list containing the
<\code{\link[carrier:crate]{crated}}> plot-functions (if \code{savePlots = TRUE}).
}
\description{
Run a random sub-sampling modification of the SDAR algorithm as
originally standardized in ASTM E3076-18. As the original version uses
numerous linear regressions (\code{.lm.fit()} from the stats-package), can be
painfully slow for test data with high resolution. The lazy variant of the
algorithm will use several random sub-samples of the data to find the best
estimate for the fit-range within the data.
}
\note{
The function can use parallel processing via the
\code{\link[furrr:furrr-package]{furrr-package}}. To use this feature, set up a plan
other than the default sequential strategy beforehand. As random values are
drawn, set a random seed (see: \code{\link[=set.seed]{set.seed()}}) beforehand to get reproducible
results.
}
\examples{
# Synthesize a test record resembling Al 6060 T66
# (Values according to Metallic Material Properties
# Development and Standardization (MMPDS) Handbook).
# Explicitly set names to "strain" and "stress".
Al_6060_T66 <- synthesize_test_data(slope = 68000,
                                    yield.y = 160,
                                    ultimate.y = 215,
                                    ultimate.x = 0.091,
                                    x.name = "strain",
                                    y.name = "stress")

# use sdarr.lazy() to analyze the synthetic test record
# (using relaxed settings for the noise-free synthetic data)
# will print a report and give a plot of the final fit
\donttest{
result <- sdarr.lazy(Al_6060_T66, strain, stress,
                     cutoff_probability = 0.8,
                     fit.rep = 2)
}

}
\references{
Lucon, E. (2019). Use and validation of the slope determination
by the analysis of residuals (SDAR) algorithm (NIST TN 2050; p. NIST TN
2050). National Institute of Standards and Technology.
https://doi.org/10.6028/NIST.TN.2050

Standard Practice for Determination of the Slope in the Linear
Region of a Test Record (ASTM E3076-18). (2018).
https://doi.org/10.1520/E3076-18

Graham, S., & Adler, M. (2011). Determining the Slope and Quality
of Fit for the Linear Part of a Test Record. Journal of Testing and
Evaluation - J TEST EVAL, 39. https://doi.org/10.1520/JTE103038
}
\seealso{
\code{\link[=sdarr]{sdarr()}} for the standard SDAR-algorithm.
}
