% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdarrLazy.R
\name{sdarr.lazy}
\alias{sdarr.lazy}
\title{sdarr.lazy}
\usage{
sdarr.lazy(
  data,
  x,
  y,
  fit.rep = 5,
  fit.candidates = 20,
  cutoff_probability = 0.975,
  quality_penalty = 0.1,
  enforce_subsampling = F,
  verbose = "report",
  showPlots = "report",
  savePlots = F
)
}
\arguments{
\item{data}{Data record to analyze. Labels of the data columns will be used
as units.}

\item{x}{Use tidy selection to specify x within the data.}

\item{y}{Use tidy selection to specify y within the data.}

\item{fit.rep}{Repetitions of random sub-sampling and fitting.}

\item{fit.candidates}{Give a number for selecting optimum fit candidates
(ordered decreasingly by normalized residuals) for each of the repetitions
of random sub-sampling.}

\item{cutoff_probability}{Cut-off probability for estimating optimum size of
sub-sampled data range via logistic regression.}

\item{quality_penalty}{Factor to down-weight fits with inferior data- and
fit-quality metrics.}

\item{enforce_subsampling}{Set to TRUE, to use sub-sampling method even when
it is computationally more expensive than the standard SDAR-algorithm.}

\item{verbose}{Give informational messages during computation defaults to
"report" to only show a summarizing information set to "all" to also give
messages from the individual steps set to "none" to be quiet. Can be
abbreviated.}

\item{showPlots}{Show plots during computation defaults to "report" to only
show the plot of the final fit set to "all" to also show plots from the
individual steps set to "none" to be quiet. Can be abbreviated.}

\item{savePlots}{Give plot functions with the result.}
}
\value{
A list containing a data-frame with the results of the final fit, a
  list with the quality- and fit-metrics, and a list containing the crated
  plot-functions (if savePlots was set to TRUE).
}
\description{
Run a random sub-sampling modification of the SDAR algorithm as
  originally standardized in ASTM E3076-18. As the original version uses
  numerous linear regressions (.lm.fit form the stats-package), can be
  painfully slow for test data with high resolution. The lazy variant of the
  algorithm will use several random sub-samples of the data to find the best
  estimate for the fit-range within the data.
}
\note{
The function will use parallel processing via the furrr-package to use
  this feature, set up a plan other than the default sequential strategy
  beforehand.
}
\references{
Lucon, E. (2019). Use and validation of the slope determination
  by the analysis of residuals (SDAR) algorithm (NIST TN 2050; p. NIST TN
  2050). National Institute of Standards and Technology.
  https://doi.org/10.6028/NIST.TN.2050

Standard Practice for Determination of the Slope in the Linear
  Region of a Test Record (ASTM E3076-18). (2018).
  https://doi.org/10.1520/E3076-18
}
