% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdarrStandard.R
\name{sdarr}
\alias{sdarr}
\title{SDAR-algorithm}
\usage{
sdarr(data, x, y, verbose = "report", showPlots = "report", savePlots = FALSE)
}
\arguments{
\item{data}{Data record to analyze. Labels of the data columns will be used
as units.}

\item{x, y}{Use <\code{\link[dplyr:dplyr_tidy_select]{tidy-selections}}> to specify x
and y within the data record.}

\item{verbose, showPlots}{Give informational messages and plots during
computation. Defaults to \code{"report"} to only show a report and a plot of the
final fit. Set to \code{"all"} to also give messages from the individual steps.
Set to \code{"none"} to be quiet. Can be abbreviated.}

\item{savePlots}{Set to \code{TRUE} to get plot-functions with the result for
later use.}
}
\value{
A list containing a data-frame with the results of the final fit,
lists with the quality- and fit-metrics, and a list containing the
<\code{\link[carrier:crate]{crated}}> plot-functions (if \code{savePlots = TRUE}).
}
\description{
Run the SDAR algorithm as standardized in ASTM E3076-18. Will
use numerous linear regressions (\code{.lm.fit()} from the stats-package) and
can be painfully slow for test data with high resolution.
}
\note{
The function can use parallel processing via the
\link[furrr:furrr-package]{furrr::furrr-package}. To use this feature, set up a plan other than the
default sequential strategy beforehand.
}
\examples{
# Synthesize a test record resembling Al 6060 T66
# (Values according to Metallic Material Properties
# Development and Standardization (MMPDS) Handbook).
# Explicitly set names to "strain" and "stress",
# set effective number of bits in the x-data to 12
# to limit the number of data points.
Al_6060_T66 <- synthesize_test_data(slope = 68000,
                                    yield.y = 160,
                                    ultimate.y = 215,
                                    ultimate.x = 0.091,
                                    x.name = "strain",
                                    y.name = "stress",
                                    enob.x = 12)

# use sdarr() to analyze the synthetic test record
# will print a report and give a plot of the final fit
result <- sdarr(Al_6060_T66, strain, stress)

}
\references{
Lucon, E. (2019). Use and validation of the slope determination
by the analysis of residuals (SDAR) algorithm (NIST TN 2050; p. NIST TN
2050). National Institute of Standards and Technology.
https://doi.org/10.6028/NIST.TN.2050

Standard Practice for Determination of the Slope in the Linear
Region of a Test Record (ASTM E3076-18). (2018).
https://doi.org/10.1520/E3076-18

Graham, S., & Adler, M. (2011). Determining the Slope and Quality
of Fit for the Linear Part of a Test Record. Journal of Testing and
Evaluation - J TEST EVAL, 39. https://doi.org/10.1520/JTE103038
}
\seealso{
\code{\link[=sdarr.lazy]{sdarr.lazy()}} for the random sub-sampling modification of the
SDAR-algorithm.
}
